Upgraded esp32-camera component back to its upstream version, also fixed an issue with IDF v5.0
Ported the rest of the project to IDF v5.0

Disabled router solicitation with the DHCP server for AP.  This will allow multi-homed devices to connect to the AP, while keeping internet access.

Updated README.md to include the instructions on how to use the AsyncAPI code generator to build the messaging schema component.